src/tests/e2e/requirements.txt
# Copyright (C) 2024-2025 Intel Corporation
# SPDX-License-Identifier: Apache-2.0

pytest==7.4.4 # MIT
allure-pytest==2.13.5 # Apache 2.0
requests==2.32.3 # Apache 2.0
kr8s==0.17.4 # BSD 3-Clause License
